+++ source/main/cResourceCount.cc
@@ -363,7 +363,7 @@ void cResourceCount::SetGradientCount(cAvidaContext& ctx, cWorld* world, const i
                       const int& halo_anchor_x, const int& halo_anchor_y, const int& move_speed, 
                       const double& plateau_inflow, const double& plateau_outflow, const int& is_plateau_common, 
                       const double& floor, const int& habitat, const int& min_size, const int& max_size,
-                      const int& config, const int& count, const double& resistance, const int& plat_val, 
+                      const int& config, const int& count, const double& resistance, const double& plat_val, 
                       const double& threshold, const int& refuge) 
 {
   assert(res_id >= 0 && res_id < resource_count.GetSize());
